Add filter component for problem list
