+++ app/c/[cid]/problem-card.tsx
@@ -83,7 +83,7 @@ export default function ProblemCard({
       href={`/c/${collection.cid}/p/${problem.pid}${searchParams}`}
       prefetch={true}
     >
-      <div className="bg-white p-6 pb-5 pr-[22px] md:p-8 md:pb-7 my-4 sm:my-6 rounded-2xl soft-shadow-xl">
+      <div className="bg-white p-6 pb-5 pr-[22px] md:p-8 md:pb-7 mb-4 sm:mb-6 rounded-2xl soft-shadow-xl">
         <div className="flex mb-2.5 md:mb-4 items-start">
           <div
             className={`min-w-1 self-stretch ${titleLineColor} rounded-full mr-2 md:mr-2.5`}
